Deploy on push, test 2

+++ dev/deploy.php
@@ -1,6 +1,6 @@
 <?php
     // Use in the "Post-Receive URLs" section of your GitHub repo.
-    if ( $_POST['payload'] ) {
+    if ( $_POST['payload'] && false ) {
         file_put_contents('log/payload.txt', print_r($_POST['payload'], FILE_APPEND);
 
         $exec_string = 'cd /var/customers/webs/jmm/weltengeschichte.de.repo/ && git reset --hard HEAD && git pull';
@@ -8,5 +8,5 @@
         file_put_contents('log/output.txt', $output, FILE_APPEND);
     }
 
-    file_put_contents('log/post.txt', print_r($_POST), FILE_APPEND);
+    file_put_contents('log/post.txt', print_r($_POST));
 ?>
